requirements.txt
matplotlib-inline==0.1.3
numpy==1.21.2
pandas==1.3.3
requests==2.26.0
SQLAlchemy==1.4.25
PyYAML==5.4.1


